Only send start / exit via console.clear when running headless
console.clear clears the console (d'uh). We use it for communication in
headless mode because it won't ever be called with an argument from normal
code - so it's a good option for communicating data via the devtools
protocol (less ambiuity than the other console methods).

In non-headless mode there's no need to communicate start / exit and we don't
want to clear the console - so let's just print out the exit code instead.