Wait for headless browser process to exit in Browser.Stop()
I've been seeing `panic: send on closed channel` errors from Run() and believe
those could happen because we close the out channel before the browser has
died and definitely won't send more messages. Let's see. Good behaivor in any
case.

As we expect to see EOF in the receive loop from the websocket on browser
shutdown and waiting for the browser to actually stop would keep the cmd != nil
longer we switch our method for detecting whether the browser was stopped to a
bool flag.