Wait a few milliseconds before exiting to await outstanding events
10 milliseconds is plenty for outstanding console events to arrive on close but
short enough to be insignificant for humans.

It's easy to accidentally swallow errors by pushing something onto the event
loop (async) but not awaiting it - even a setTimeout 0 in js would solve this
but it's a burden on the developer to think of this. This hack makes things a
little more user friendly. 10 milliseconds seems random - and it
is. trial and error showed 1 would be enough - but 10 is a nice round number
that gives us some headroom.

Note that those errors are only swallowed in the terminal output - opening the
served index.html shows them without this hack because we're not closing the
window.