Refactor Event.Args resolution
Until now we just used the value property of each argument and relied on
overriding console.log json stringifying the arguments to receive useful
non-primitive values. This had multiple problems:
- iframes don't use the parent windows console methods - i.e. no json
  stringification on console.log inside an iframe
- looking at log output in the browser unnecessarily destroyed information
  about the logged values by turning them into strings rather than logging the
  actual references. This could have also been prevented by checking whether
  the script is being run headless - but this feels cleaner

It's not perfect - non-primitives are only logged as previews (e.g. no
resolution of nested non-primitives). But I don't have a need for that for now
so yagni.