Refactor script loading: Simplify and allow multiple script modules
Running multiple scripts is a common use case and wasn't possible before -
as flags don't take multiple arguments by default it's easier to refactor
the binary to take files as remaining args and window.args as a flag. While
we're at it we can introduce a listen address flag as well.

The splitting into servePath + filename is not necessary if we assume we serve
the current working directory and (1) any script files are relative to
that and (2) fetch relative to their import.meta.url
(e.g. `new URL(relativePath, import.meta.url)`). `import` automatically does
that and i guess there's no oher way if we want to suppor multiple script files
that are not necessarily in the same directory.

Also no need to log start not called warning - just error after a
second.