Only enable domain if not already enabled and simplify Unsubscribe lock
Not all domains have a no-op enable method - e.g. Fetch.enable takes params and
we want to enable it manually via Execute. Until now we re-enabled because
that's not a problem for no-op enables. Don't do that anymore.

As the Domain.disable check loops the p.events map synchronously and can't
block there's no need to have it in a separate lock - so let's simplify that as
well.