Add RunOn to allow scraping using iframes (trick same origin policy)
I like to scrape web pages using devtools - when I need to navigate across
pages using an iframe allows keeping state. As browsers prevent access to the
document of cross-origin iframes we need to trick the browser into thinking our
code comes from the domain we want to scrape. Hackernews tells me this is how
selenium worked until they abandoned it due to shortcomings (i guess because
some things can't be done from js - e.g. activating css hover).

As some pages prevent loading in iframes via headers we'll proabably have to
intercept everything on the domain of interest - but let's leave that out for
now as not a lot of pages bother to go that far.