Rewrite: Move logic into js and simplify go client. Drop RunOn.
TLDR: Completly changed architecture due to requirements changing to support
cross origin access for scraping / automation. Moved a huge amount of logic
into js because it's nicer to work in / reusable outside of go.

RunOn proved to be not as helpful as expected - while it circumvents origin
isolation for the top origin it breaks down for nested iframes with differing
origins. More and more sites have those to ask for consent to fulfill GDPR
requirements. The consent checks are often done in cross origin
iframes (CMPs).

I spent some time looking into getting universal access (i.e. cross origin
access) without disabling web security completely (which would be a really bad
idea when scraping - who knows what shady scripts one runs across in the wild
wild web). Page.createIsolatedWorld({grantUniversalAccess}) +
Runtime.evaluate() [1] looked like the perfect fit, but sadly it's broken right
now [2] and I got tired of jumping through the chromium source without any hope
of actually compiling chromium / getting debug symbols for gdb [3]. It doesn't
work in firefox either [4]

With no hope for universal access that out of the window the next best thing is
allowing js to control the browser (yo dawg). The go wrapper should be as small
as possible as it's only in go by necessity (i want to use it from other go
code bases) - and without generating types working with the devtools protocol
is just too verbose in go. And why shouldn't js have access? It's running on an
isolated origin (so no danger) and ES6 makes working with async and nested
types (destructuring) a breeze.

So here we are. I just threw out pretty much everything and redesigned the API
- the perks of a tiny code base with no external users. If only the real world
was like that...

In the end, I like the current public API a lot more as well - the previous
iterations felt hacky (I'm still getting the hang of actually writing async
code in go - never had a reason to add that complexity before) and this feels
clean. API and tests are getting neglected because I'm not convinced the
architecture is stable yet. But hey, it works - enough of this for now and on
to more fun things.

[1]
universalAccess allows access to all other origins - the
`--disable-web-security` flag sets it as well, but apparently in more places.
`universalAccessFromFileUrls` would have been another good candidate but that one
seems to be restricted to chromium for android.
- https://source.chromium.org/search?q=UniversalAccessFromFileUrls
- https://chromedevtools.github.io/devtools-protocol/tot/Page/#method-createIsolatedWorld
- https://chromedevtools.github.io/devtools-protocol/tot/Runtime/#method-evaluate

[2] https://stackoverflow.com/questions/65513720/page-createisolatedworld-grantuniveralaccess-flag-does-not-grant-universal-acces

[3] The build requirements for chromium are crazy. 100GB hdd, 16GB ram, 10h+ on
recent hardware. Far far out of the possibilities of my trusty old notebook -
and even disregarding that there's a huge amount of hoops to jump through to
actually commit something compared to the open source projects i normally
contribute to.

[4] https://bugzilla.mozilla.org/show_bug.cgi?id=1597877

https://chromium.googlesource.com/chromium/src/+/master/docs/linux/build_instructions.md