Refactor test to use json fixture
Keeping the expected results updated is tiresome - especially with line and
column numbers of error stack traces changing with any change to run.html.

The index is added to the key to keep the order in results.json in line with
the test cases (json.Marshall orders maps alphabetically) - using a slice would
also work but makes lookup more tiresome.