Refactor html util: Import files sync
Loading the files as multiple script tags causes the files to be loaded
independently of each other and therefore not executed together
synchronously. Doing all the imports in a single module makes the execution
synchronous. This can be helpful for setup as it ensures any async execution
inside those files happens after all files have been loaded.

We lose information about which file failed to load but that is easy to get by
running with gui and checking the console. As I don't expect this to be a major
annoyance and I'd have to rewrite my shitty js test library which assumes async
== all test files have been loaded I'm going with this solution.