Rename Runner to H
It should have been Headless - but just like the go test package exports
testing.T instead of testing.Test we'll go with a short name. Being verbose is
not a virtue.

As the reference name h shows this it was originally called Headless and i just
renamed it to Runner to prevent the above. I like this way more.