Generalize html
Redefining close to console.clear is an assumption that got removed in the rest
of the codebase (clear messages don't end the run anymore). Opening iframes and
the rest of the set up js is also probably not as general so I removed those as
well for now. Specification should be possible at the point of use - headless
is just a library and should provide ways to solve common use cases but not
introduce too many assumptions.