Do not set up headless connection when called non-programmatically
When using headless as a test runner a common use case is running the tests in
manually in another browser (i.e. not H.Browser). When opening the HTML
manually we want to replicate what headless does but with the ability to watch
execution / debug / whatever - i.e. we want the same html headless opens. We
also don't want to connect headless because we don't want messages to be send
to the backend. As long as test runner use cases don't touch the exposed
headless instance this should simply work. And touching the headless instance
isn't required for the test runner use case afaict - everything is on our
origin and we control all the code - iframes should be enough. Let's see.

We have to move the template to be able to access it in the setup script - but
that doesn't change anything because it's a template and inert until inserted.