Reduce impact of method call deadlock
The idea behind the channel was to serialize execution - sadly having a buffer
of 2 is not enough; if a handler communicates with the browser it can end up
deadlocked. The following

    readFile("foo.txt")
    console.log(1)
    console.log(2)
    console.log(3)

will get stuck because readFile will send the result to the browser using eval
and block until it receives the eval result. log 1 and log 2 will fill the
channel buffer and the write of log 3 to the channel will block the rw loop -
and thus prevent readFile from ever receiving the eval result and finishing.

Panicing instead of deadlocking will remove the pain of diagnosing the problem
when it happens again - for now let's increase the buffer to some big-ish
number; maybe that's enough - if not we can always look into building our own
infinite buffered channel (removing serialization is not something i want -
while awaiting is necessary for bound fns to get results e.g. console.log is
not normally awaited and "console.log(1); window.close(1);" would race and
potentially swallow the log without serialization)