Replace setting base url in template url with nav for headless main
The headless main cmd needs to execute the code from TemplateHTML it loads via
setDocumentContent from a domain - the default "about:blank" is not enough to
e.g. get relative links to be parsed correctly by DOMParser. base solves that
but also breaks CORS when the template html is actually served instead - the
actual serving domain is replaced and same origin requests to that same domain
fail. doing a navigation to a safe non-existant url (.invalid urls are
guaranteed to not exist) solves the same problem in a more localized way.